requirements.txt
click>=8.0.0
pyyaml>=5.1 